Remove default control manager, rename esql selections manager
